LayoutTests/fast/dom/HTMLAnchorElement/anchor-download-synthetic-click-expected.txt

CONSOLE MESSAGE: line 22: Non user-triggered activations of anchors that have a download attribute are ignored.
Test that synthetic clicks on an anchor with a download attribute are ignored.

This test passes if you do not see any 'Download started' message above.

Blob URL
